requirement.txt
asgiref==3.5.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.4
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
typing-extensions==4.3.0
urllib3==1.26.12